+++ src/client/components/AuthProvider.jsx
@@ -1,0 +1,7 @@
+import {createContext, useContext, useMemo} from "react";
+
+const AuthProvider = ({children}) => {
+
+}
+
+export default AuthProvider;
