+++ src/client/components/Login.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Login = () => {
+    <div>
+        <h1>This is the login page</h1>
+    </div>
+}
