+++ src/client/components/NavBar.jsx
